+++ src/renderer/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="monetization" content="$ilp.uphold.com/rEgHHBQGDUDP" />
+    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#403f4d" />
+    <meta name="description" content="todometer: a meter-based to-do list" />
+    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo128.png" />
+    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <title>todometer</title>
+    <style>
+      #loading {
+        margin: 40px auto;
+        font-family: "Helvetica", "Arial", sans-serif;
+        font-weight: lighter;
+        color: #fbfafb;
+        text-align: center;
+      }
+    </style>
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root">
+      <div id="loading">loading...</div>
+    </div>
+    <script src="src/index.jsx" type="module"></script>
+  </body>
+</html>
