+++ src/renderer/robots.txt
@@ -1,0 +1,2 @@
+# https://www.robotstxt.org/robotstxt.html
+User-agent: *
