+++ src/renderer/src/App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import TodoDate from "./components/TodoDate";
-import ItemList from "./components/ItemList";
-import { AppStateProvider } from "./AppContext";
+import TodoDate from "./components/TodoDate.jsx";
+import ItemList from "./components/ItemList.jsx";
+import { AppStateProvider } from "./AppContext.jsx";
 
 function App() {
   return (
