+++ src/renderer/src/App.test.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "@testing-library/react";
 import { format } from "date-fns";
-import App from "./App";
+import App from "./App.jsx";
 
 test("renders the day of the month", () => {
   const { getByText } = render(<App />);
