+++ src/renderer/src/AppContext.jsx
@@ -1,5 +1,5 @@
 import React, { createContext, useContext, useReducer } from "react";
-import { loadState, saveState } from "./local-storage";
+import { loadState, saveState } from "./local-storage.js";
 import { format } from "date-fns";
 
 export const AppContext = createContext();
