+++ src/renderer/src/components/AddItemForm.jsx
@@ -1,5 +1,5 @@
 import React, { useRef } from "react";
-import { useAppReducer } from "../AppContext";
+import { useAppReducer } from "../AppContext.jsx";
 
 import styles from "./AddItemForm.module.scss";
