+++ src/renderer/src/components/Progress.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useItems, useAppState } from "../AppContext";
+import { useItems, useAppState } from "../AppContext.jsx";
 
 import styles from "./Progress.module.scss";
