+++ src/renderer/src/index.scss
@@ -1,4 +1,4 @@
-@import "./styles/variables";
+@import "styles/variables";
 
 body {
   margin: 0;
