+++ src/renderer/src/styles/variables.scss
